Add a min-width tag to the leaderboard
The leaderboard currently has some overlapping elements when using a "portrait-style" screen, this is due to the floating nature of the divs used on the website. The easiest solution right now would be to introduce a min-width and force browsers to apply a scrollbar so that all content will be visible.

Signed-off-by: Rens Rikkerink <ikkerens@ikkerens.rat.fuelrats.com>